test: split component tests into per-file test suites
Move TokenUsageBar and RetryBadge tests out of BudgetGauge.test.tsx into
their own dedicated test files; add extra coverage cases for each component.
